tried to seeparate the sidebar into a separate file, but i keep getting compiler errors. moved the small things into files and folders.
