+++ iPad browser/Extensions/ContentViewColor.swift
@@ -1,0 +1,21 @@
+//
+//  ContentViewColor.swift
+//  Aura
+//
+//  Created by Quinn O'Donnell on 4/27/24.
+//
+
+import Foundation
+import SwiftUI
+
+extension UIColor {
+    func toHex() -> String? {
+        guard let components = cgColor.components, components.count >= 3 else {
+            return nil
+        }
+        let r = Float(components[0])
+        let g = Float(components[1])
+        let b = Float(components[2])
+        return String(format: "%02lX%02lX%02lX", lroundf(r * 255), lroundf(g * 255), lroundf(b * 255))
+    }
+}
