+++ iPad browser/Helpers/DropViewDelegate.swift
