+++ iPad browser/Helpers/IconsPicker.swift
