+++ iPad browser/Helpers/Long Variables.swift
