+++ iPad browser/Helpers/Suggestions/SuggestionsData.swift
