+++ iPad browser/Helpers/Suggestions/SuggestionsModel.swift
@@ -1,0 +1,14 @@
+//
+//  SuggestionsModel.swift
+//  Aura
+//
+//  Created by Quinn O'Donnell on 4/27/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct Suggestion: Identifiable, Codable {
+    var id: String
+    var url: String
+}
