+++ iPad browser/Helpers/Suggestions/WebsiteSuggestion.swift
@@ -1,0 +1,15 @@
+//
+//  WebsiteSuggestion.swift
+//  Aura
+//
+//  Created by Quinn O'Donnell on 4/27/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct WebsiteSuggestion: Codable {
+    var url: String
+    var title: String
+}
+
