+++ iPad browser/Helpers/urlFormatter.swift
